+++ src/bootstrap/bootstrap.js
@@ -129,3 +129,8 @@ class TextDecoder {
 
 globalThis.TextEncoder = TextEncoder;
 globalThis.TextDecoder = TextDecoder;
+
+// === Reexports === //
+
+globalThis.structuredClone = Deno.core.ops.op_structured_clone;
+globalThis.queueMicrotask = Deno.core.ops.op_queue_microtask;